Extract CopyButton into separate component
